docs: sd-video docs [skip chromatic] (#1331)
